Macro dslab_core::cast

macro_rules! cast {
    ( match $event:ident.data { $( $type:ident { $($tt:tt)* } => { $($expr:tt)* } )+ } ) => { ... };
}

Enables the use of pattern matching syntax for processing different types of events by downcasting the event payload from EventData to user-defined types.

Note that match arms need not be exhaustive. However, if the event payload does not match any of specified arms, the macro will log the event as unhandled under ERROR level.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

#[derive(Clone, Serialize)]
pub struct AnotherEvent {
    another_field: f64,
}

pub struct Component {
    ctx: SimulationContext,
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                // some event processing logic...
            }
            AnotherEvent { another_field } => {
                // some event processing logic...
            }
        })

   }
}

let mut sim = Simulation::new(123);
let mut comp_ctx = sim.create_context("comp");
let comp_id = sim.add_handler("comp", Rc::new(RefCell::new(Component { ctx: comp_ctx })));
let client_ctx = sim.create_context("client");
client_ctx.emit(SomeEvent{ some_field: 16 }, comp_id, 1.2);
client_ctx.emit(AnotherEvent{ another_field: 1.6 }, comp_id, 2.5);
sim.step_until_no_events();