Struct dslab_core::context::SimulationContext

pub struct SimulationContext { /* private fields */ }

A facade for accessing the simulation state and producing events from simulation components.

Implementations

impl SimulationContext

pub fn id(&self) -> Id

Returns the identifier of component associated with this context.

Examples

use dslab_core::{Simulation, SimulationContext};

let mut sim = Simulation::new(123);
let comp_ctx = sim.create_context("comp");
let comp_id = comp_ctx.id();
assert_eq!(comp_id, 0); // component ids are assigned sequentially starting from 0

pub fn name(&self) -> &str

Returns the name of component associated with this context.

Examples

use dslab_core::{Simulation, SimulationContext};

let mut sim = Simulation::new(123);
let comp_ctx = sim.create_context("comp");
let comp_name = comp_ctx.name();
assert_eq!(comp_name, "comp");

pub fn time(&self) -> f64

Returns the current simulation time.

Examples

use dslab_core::{Simulation, SimulationContext};

let mut sim = Simulation::new(123);
let comp_ctx = sim.create_context("comp");
let time = comp_ctx.time();
assert_eq!(time, 0.0);

pub fn rand(&self) -> f64

Returns a random float in the range [0, 1) using the simulation-wide random number generator.

Examples

use dslab_core::{Simulation, SimulationContext};

let mut sim = Simulation::new(123);
let mut comp_ctx = sim.create_context("comp");
let f: f64 = comp_ctx.rand();
assert!(f >= 0.0 && f < 1.0);

pub fn gen_range<T, R>(&self, range: R) -> T
where T: SampleUniform, R: SampleRange<T>,

Returns a random number in the specified range using the simulation-wide random number generator.

Examples

use dslab_core::{Simulation, SimulationContext};

let mut sim = Simulation::new(123);
let mut comp_ctx = sim.create_context("comp");
let n: u32 = comp_ctx.gen_range(1..=10);
assert!(n >= 1 && n <= 10);
let f: f64 = comp_ctx.gen_range(0.1..0.5);
assert!(f >= 0.1 && f < 0.5);

pub fn sample_from_distribution<T, Dist: Distribution<T>>( &self, dist: &Dist ) -> T

Returns a random value from the specified distribution using the simulation-wide random number generator.

pub fn random_string(&self, len: usize) -> String

Returns a random alphanumeric string of specified length using the simulation-wide random number generator.

pub fn emit<T>(&self, data: T, dst: Id, delay: f64) -> EventId
where T: EventData,

Creates new event with specified payload, destination and delay, returns event id.

The event time will be current_time + delay. It is not allowed to create events before the current simulation time, so delay should be non-negative.

The event source will be equal to id. See emit_as() if you want to emit event on behalf of some other component.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

pub struct Component {
    ctx: SimulationContext,
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                assert_eq!(self.ctx.time(), 1.2);
                assert_eq!(event.time, 1.2);
                assert_eq!(event.id, 0);
                assert_eq!(some_field, 16);
            }
        })

   }
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
let comp2_id = sim.add_handler("comp2", Rc::new(RefCell::new(Component { ctx: comp2_ctx })));
let event_id = comp1_ctx.emit(SomeEvent{ some_field: 16 }, comp2_id, 1.2);
assert_eq!(event_id, 0); // events ids are assigned sequentially starting from 0
sim.step();
assert_eq!(sim.time(), 1.2);

use serde::Serialize;
use dslab_core::{Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), -1.0); // will panic because of negative delay

pub fn emit_ordered<T>(&self, data: T, dst: Id, delay: f64) -> EventId
where T: EventData,

This and all other emit_ordered... functions are special variants of normal emit_... functions that allow adding events to ordered event deque instead of heap, which may improve simulation performance.

Ordered events should be emitted in non-decreasing order of their time, otherwise the simulation will panic.

Examples

use serde::Serialize;
use dslab_core::{Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 1.0);
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 1.0);
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 2.0);
sim.step();
assert_eq!(sim.time(), 1.0);
sim.step();
assert_eq!(sim.time(), 1.0);
sim.step();
assert_eq!(sim.time(), 2.0);

use serde::Serialize;
use dslab_core::{Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 2.0);
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 1.0); // will panic because of broken time order

pub fn can_emit_ordered(&self, delay: f64) -> bool

Checks whether it is safe to emit an ordered event with the specified delay.

The time of new event must be not less than the time of the previously emitted ordered event.

Returns true if this condition holds and false otherwise.

Examples

use serde::Serialize;
use dslab_core::{Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 1.0);
assert!(comp1_ctx.can_emit_ordered(1.0)); // 1.0 == 1.0
assert!(comp1_ctx.can_emit_ordered(1.1)); // 1.1 > 1.0
assert!(!comp1_ctx.can_emit_ordered(0.9)); // 0.9 < 1.0
comp1_ctx.emit_ordered(SomeEvent{}, comp2_ctx.id(), 1.5);
assert!(!comp1_ctx.can_emit_ordered(1.0)); // 1.0 < 1.5
sim.step();
assert_eq!(sim.time(), 1.0);
assert!(comp1_ctx.can_emit_ordered(1.0)); // 2.0 > 1.5
assert!(!comp1_ctx.can_emit_ordered(0.3)); // 1.3 < 1.5

pub fn emit_now<T>(&self, data: T, dst: Id) -> EventId
where T: EventData,

Creates new immediate (zero-delay) event with specified payload and destination, returns event id.

This is a shorthand for emit() with zero delay.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

pub struct Component {
    ctx: SimulationContext,
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                assert_eq!(self.ctx.time(), 0.0);
                assert_eq!(event.time, 0.0);
                assert_eq!(event.id, 0);
                assert_eq!(some_field, 16);
            }
        })

   }
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
let comp2_id = sim.add_handler("comp2", Rc::new(RefCell::new(Component { ctx: comp2_ctx })));
let event_id = comp1_ctx.emit_now(SomeEvent{ some_field: 16 }, comp2_id);
assert_eq!(event_id, 0); // events ids are assigned sequentially starting from 0
sim.step();
assert_eq!(sim.time(), 0.0);

pub fn emit_ordered_now<T>(&self, data: T, dst: Id) -> EventId
where T: EventData,

See Self::emit_ordered.

pub fn emit_self<T>(&self, data: T, delay: f64) -> EventId
where T: EventData,

Creates new event for itself with specified payload and delay, returns event id.

This is a shorthand for emit() with event destination equals id.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

pub struct Component {
    ctx: SimulationContext,
}

impl Component {
    fn start(&mut self) {
        self.ctx.emit_self(SomeEvent{ some_field: 16 }, 6.4);
    }
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                assert_eq!(self.ctx.time(), 6.4);
                assert_eq!(event.time, 6.4);
                assert_eq!(event.id, 0);
                assert_eq!(event.src, self.ctx.id());
                assert_eq!(some_field, 16);
            }
        })

   }
}

let mut sim = Simulation::new(123);
let comp1 = Rc::new(RefCell::new(Component { ctx: sim.create_context("comp1") }));
let comp1_id = sim.add_handler("comp1", comp1.clone());
comp1.borrow_mut().start();
sim.step();
assert_eq!(sim.time(), 6.4);

pub fn emit_ordered_self<T>(&self, data: T, delay: f64) -> EventId
where T: EventData,

See Self::emit_ordered.

pub fn emit_self_now<T>(&self, data: T) -> EventId
where T: EventData,

Creates new immediate event for itself with specified payload, returns event id.

This is a shorthand for emit() with event destination equals id and zero delay.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

pub struct Component {
    ctx: SimulationContext,
}

impl Component {
    fn start(&mut self) {
        self.ctx.emit_self_now(SomeEvent{ some_field: 16 });
    }
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                assert_eq!(self.ctx.time(), 0.0);
                assert_eq!(event.time, 0.0);
                assert_eq!(event.id, 0);
                assert_eq!(event.src, self.ctx.id());
                assert_eq!(some_field, 16);
            }
        })

   }
}

let mut sim = Simulation::new(123);
let comp1 = Rc::new(RefCell::new(Component { ctx: sim.create_context("comp1") }));
let comp1_id = sim.add_handler("comp1", comp1.clone());
comp1.borrow_mut().start();
sim.step();
assert_eq!(sim.time(), 0.0);

pub fn emit_ordered_self_now<T>(&self, data: T) -> EventId
where T: EventData,

See Self::emit_ordered.

pub fn emit_as<T>(&self, data: T, src: Id, dst: Id, delay: f64) -> EventId
where T: EventData,

Creates new event with specified payload, source, destination and delay, returns event id.

This is an extended version of emit() for special cases when the event should be emitted on behalf of another component.

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
    some_field: u32,
}

pub struct Component {
    ctx: SimulationContext,
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { some_field } => {
                assert_eq!(self.ctx.time(), 2.4);
                assert_eq!(event.time, 2.4);
                assert_eq!(event.id, 0);
                assert_eq!(event.src, 0);
                assert_eq!(self.ctx.id(), 1);
                assert_eq!(some_field, 8);
            }
        })

   }
}

let mut sim = Simulation::new(123);
let comp1 = Rc::new(RefCell::new(Component { ctx: sim.create_context("comp1") }));
let comp1_id = sim.add_handler("comp1", comp1);
let comp2 = Rc::new(RefCell::new(Component { ctx: sim.create_context("comp2") }));
let comp2_id = sim.add_handler("comp2", comp2);
let mut other_ctx = sim.create_context("other");
other_ctx.emit_as(SomeEvent{ some_field: 8 }, comp1_id, comp2_id, 2.4);
sim.step();
assert_eq!(sim.time(), 2.4);

pub fn emit_ordered_as<T>( &self, data: T, src: Id, dst: Id, delay: f64 ) -> EventId
where T: EventData,

See Self::emit_ordered.

pub fn cancel_event(&self, id: EventId)

Cancels the specified event.

Use EventId obtained when creating the event to cancel it. Note that already processed events cannot be cancelled.

Examples

use serde::Serialize;
use dslab_core::{Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
let event1 = comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), 1.0);
let event2 = comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), 2.0);
sim.step();
comp1_ctx.cancel_event(event2);
sim.step_until_no_events();
assert_eq!(sim.time(), 1.0);

pub fn cancel_events<F>(&self, pred: F)
where F: Fn(&Event) -> bool,

Cancels events that satisfy the given predicate function.

Note that already processed events cannot be cancelled.

Examples

use serde::Serialize;
use dslab_core::{Event, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
let event1 = comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), 1.0);
let event2 = comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), 2.0);
let event2 = comp1_ctx.emit(SomeEvent{}, comp2_ctx.id(), 3.0);
comp1_ctx.cancel_events(|e| e.id < 2);
sim.step();
assert_eq!(sim.time(), 3.0);

pub fn cancel_heap_events<F>(&self, pred: F)
where F: Fn(&Event) -> bool,

Same as Self::cancel_events, but ignores events added through emit_ordered_... methods.

pub fn lookup_name(&self, id: Id) -> String

Returns component name by its identifier.

Examples

use std::cell::RefCell;
use std::rc::Rc;
use serde::Serialize;
use dslab_core::{cast, Event, EventHandler, Simulation, SimulationContext};

#[derive(Clone, Serialize)]
pub struct SomeEvent {
}

pub struct Component {
    ctx: SimulationContext,
}

impl EventHandler for Component {
    fn on(&mut self, event: Event) {
        cast!(match event.data {
            SomeEvent { } => {
                // look up the name of event source
                let src_name = self.ctx.lookup_name(event.src);
                assert_eq!(src_name, "comp1");
            }
        })

   }
}

let mut sim = Simulation::new(123);
let mut comp1_ctx = sim.create_context("comp1");
let mut comp2_ctx = sim.create_context("comp2");
let comp2_id = sim.add_handler("comp2", Rc::new(RefCell::new(Component { ctx: comp2_ctx })));
comp1_ctx.emit(SomeEvent{}, comp2_id, 1.0);
sim.step();